refactor: replace UserCreateModal and UserEditModal with success modals and remove old modal components
